+++ types/font-carrier/index.d.ts
@@ -5,6 +5,8 @@
 
 /// <reference types="node" />
 
+import { Base } from './base';
+
 export type FontType = 'ttf' | 'eot' | 'woff' | 'woff2' | 'svg';
 
 export interface FileOptions {
@@ -12,20 +14,6 @@ export interface FileOptions {
     path?: string | Buffer;
     /** 导出的字体类型 */
     types?: FontType[];
-}
-
-export class Base<T> {
-    /** 构造参数 */
-    options: T;
-    /** 设置构造参数 */
-    setOptions(options: T): void;
-    defaultOptions: T;
-    /** 获取构造参数的属性  */
-    get(key: string): string | number;
-    /** 设置构造参数的属性 */
-    set(key: string, value: string | number): void;
-    init(options: T): void;
-    constructor(arg0: T);
 }
 
 export interface FontOptions {
